Support bounded 1 GiB Git streams with durable host geometry
Keep live memory at 88 MiB and physical objects at 1 MiB while bounding
streamed blobs at 1 GiB and input packs at 1 GiB plus 16 MiB framing.
Retain small structural and buffered API limits. Use checked u64 cumulative
work and transfer accounting on WASI, with finite pass-derived call limits.
Freeze host profile V1 at 2080 references and reopen legacy logs with their
original options. Oversized URI blobs use the ordinary streamed response.

Validation: full make check (format, lint, native tests, WASIp2 checks), plus
focused host lint after extending the opt-in maintenance fixture. Actual
local loopback MinIO with ordinary Spin passed incompressible 50 MiB and
1 GiB push, cold clone, fresh fetch and Git fsck for SHA-1 and SHA-256.
1 GiB push/clone/fetch seconds: SHA-1 23.293/16.656/24.856;
SHA-256 17.743/11.923/26.742. These are local provider observations, not
remote store benchmarks. Newly added maintenance lifecycle awaits the
separate streaming compaction implementation before qualification.

+++ crates/object-log-git/src/graph.rs
@@ -627,7 +627,7 @@ mod tests {
                 .requests,
             0
         );
-        assert!(repository.operation.work_bytes() > OBJECTS * size_of::<Node>());
+        assert!(repository.operation.work_bytes() > (OBJECTS * size_of::<Node>()) as u64);
         // The failed attempt's work is retained; a retry never admits a fresh budget.
         let remaining = pack::budget::WORK_BYTES - repository.operation.work_bytes();
         repository.operation.work(remaining)?;
